guess the number
It generates a random number and the player needs to guess what that number is. After each guess, the program tells the player whether the guessed number is too large or too small until the player guesses correctly. The program also records the number of guesses made by the player, and outputs the number of guesses when the player guesses correctly. This program uses the random number generator function and loop statement in Python