Guess the Number Game
This program implements a simple number guessing game. The program will randomly generate a number between 1 and 100, and the player needs to guess this number in the process of continuous trial and error. After each guess, the program will give a hint, telling the player whether the guessed number is too high or too low. Once the player has guessed the number, the program tells the player how many times they have guessed, and the game ends.