Calculates the average of the input numbers
When the user runs this script, it will prompt the user for numbers until the user enters "Q". The script ignores invalid input and adds valid inputs to a list. If the user doesn't enter any numbers, the script will output a message that no numbers were entered. Otherwise, the script calculates the average, maximum, and minimum of the input numbers and prints them to the screen.