Create data_scraper.py
The function uses the requests library and the BeautifulSoup library to get tabular data from the web page, convert it into a Pandas DataFrame object, and perform some simple data cleaning and analysis. In the main program, it asks the user to enter the URL where the data is to be fetched and the name of the CSV file where the data is to be saved, and passes it to the "scrape_data" function.