Create data_analysis.py
This script assumes that there is a CSV file called "data.csv" that contains two columns of data, date and sales. The script will use the Pandas library to load the data from the CSV file and do some data cleaning and analysis, such as converting the date column to a Datetime object, converting the sales column to float type, calculating the sum of sales per month and year, etc. Then, the script uses the Matplotlib library to visualize the analysis results, drawing line charts and column charts of monthly sales and annual sales, respectively.